[Change] Simple refactor and Javadoc (#289)
* compilation and sonar

* forgotten some methods

* package simplification and Javadoc

* forgot tests
