+++ src/main/java/org/openstreetmap/atlas/geography/atlas/bloated/BloatedEntity.java
@@ -1,6 +1,9 @@
 package org.openstreetmap.atlas.geography.atlas.bloated;
 
 /**
+ * Simple interface for all the Bloated entities. As each one extends its parent class already
+ * (Node, Edge, Area, ...) this cannot be an abstract class.
+ *
  * @author matthieun
  */
 public interface BloatedEntity
