+++ src/main/java/org/openstreetmap/atlas/geography/atlas/bloated/BloatedPoint.java
@@ -7,12 +7,11 @@ import java.util.stream.Collectors;
 import org.openstreetmap.atlas.exception.CoreException;
 import org.openstreetmap.atlas.geography.Location;
 import org.openstreetmap.atlas.geography.Rectangle;
-import org.openstreetmap.atlas.geography.atlas.items.Edge;
 import org.openstreetmap.atlas.geography.atlas.items.Point;
 import org.openstreetmap.atlas.geography.atlas.items.Relation;
 
 /**
- * Independent {@link Edge} that contains its own data. At scale, use at your own risk.
+ * Independent {@link Point} that contains its own data. At scale, use at your own risk.
  *
  * @author matthieun
  */
