+++ src/main/java/org/openstreetmap/atlas/geography/atlas/change/ChangeAtlas.java
@@ -1,5 +1,6 @@
 package org.openstreetmap.atlas.geography.atlas.change;
 
+import java.util.Objects;
 import java.util.Optional;
 import java.util.function.BiFunction;
 import java.util.function.Function;
@@ -12,8 +13,6 @@ import org.openstreetmap.atlas.geography.atlas.AbstractAtlas;
 import org.openstreetmap.atlas.geography.atlas.Atlas;
 import org.openstreetmap.atlas.geography.atlas.AtlasMetaData;
 import org.openstreetmap.atlas.geography.atlas.builder.AtlasSize;
-import org.openstreetmap.atlas.geography.atlas.change.rule.ChangeType;
-import org.openstreetmap.atlas.geography.atlas.change.rule.FeatureChange;
 import org.openstreetmap.atlas.geography.atlas.items.Area;
 import org.openstreetmap.atlas.geography.atlas.items.AtlasEntity;
 import org.openstreetmap.atlas.geography.atlas.items.Edge;
@@ -290,12 +289,12 @@ public class ChangeAtlas extends AbstractAtlas // NOSONAR
                                 && featureChange.getChangeType() == ChangeType.ADD)
                         .map(featureChange -> entityForIdentifier
                                 .apply(featureChange.getIdentifier()))
-                        .filter(entity -> entity != null).collect(Collectors.toList()),
+                        .filter(Objects::nonNull).collect(Collectors.toList()),
                 Iterables.stream(sourceEntities)
                         .filter(entity -> !this.change.changeFor(itemType, entity.getIdentifier())
                                 .isPresent())
                         .map(entity -> entityForIdentifier.apply(entity.getIdentifier()))
-                        .filter(entity -> entity != null).collect());
+                        .filter(Objects::nonNull).collect());
     }
 
     /**
