+++ src/main/java/org/openstreetmap/atlas/geography/atlas/change/ChangeEdge.java
@@ -11,6 +11,13 @@ import org.openstreetmap.atlas.geography.atlas.items.Node;
 import org.openstreetmap.atlas.geography.atlas.items.Relation;
 
 /**
+ * {@link Edge} that references a {@link ChangeAtlas}. That {@link Edge} makes sure that all the
+ * connected {@link Node}s are {@link ChangeNode}s, and that all the parent {@link Relation}s are
+ * {@link ChangeRelation}s.
+ * <p>
+ * NOSONAR here to avoid "Subclasses that add fields should override "equals" (squid:S2160)". Here
+ * the equals from the parent works.
+ *
  * @author matthieun
  */
 public class ChangeEdge extends Edge // NOSONAR
