+++ src/main/java/org/openstreetmap/atlas/geography/atlas/change/ChangeNode.java
@@ -13,6 +13,13 @@ import org.openstreetmap.atlas.geography.atlas.items.Node;
 import org.openstreetmap.atlas.geography.atlas.items.Relation;
 
 /**
+ * {@link Node} that references a {@link ChangeAtlas}. That {@link Node} makes sure that all the
+ * connected {@link Edge}s are {@link ChangeEdge}s, and that all the parent {@link Relation}s are
+ * {@link ChangeRelation}s.
+ * <p>
+ * NOSONAR here to avoid "Subclasses that add fields should override "equals" (squid:S2160)". Here
+ * the equals from the parent works.
+ *
  * @author matthieun
  */
 public class ChangeNode extends Node // NOSONAR
