+++ src/main/java/org/openstreetmap/atlas/geography/atlas/change/ChangeType.java
@@ -1,4 +1,4 @@
-package org.openstreetmap.atlas.geography.atlas.change.rule;
+package org.openstreetmap.atlas.geography.atlas.change;
 
 /**
  * Types of changes. Here, MODIFY is omitted, as it can be translated to an ADD, for simplicity.
